Removing the setup class.
