--- includes/classes/class-setup.php
@@ -1,55 +1,0 @@
-<?php
-namespace LSX\Blocks\Classes;
-
-/**
- * @package   LSX\Blocks\Classes
- * @author    LightSpeed
- * @license   GPL-3.0+
- * @link
- * @copyright 2023 LightSpeed
- */
-
-/**
- * Setup plugin class.
- *
- * @package LSX\Blocks\Classes
- * @author  LightSpeed
- */
-class Setup {
-
-	/**
-	 * Initialize the plugin by setting localization, filters, and administration functions.
-	 */
-	public function __construct() {
-	}
-
-	/**
-	 * Loads the actions and filters.
-	 */
-	public function init() {
-		add_action( 'plugins_loaded', array( $this, 'blocks_loader' ) );
-		add_action( 'enqueue_block_editor_assets', array( $this, 'block_editor_assets' ), 5 );
-	}
-
-	/**
-	 * Load the plugin textdomain
-	 */
-	public function blocks_init() {
-		load_plugin_textdomain( 'lsx-blocks', false, basename( dirname( __FILE__ ) ) . '/languages' );
-	}
-
-	/**
-	 * Enqueue assets for backend editor
-	 *
-	 * @since 1.0.0
-	 */
-	public function block_editor_assets() {
-		// Load the compiled blocks into the editor.
-		wp_enqueue_script(
-			'lsx-blocks-block-js',
-			LSX_BLOCKS_URL . 'dist/blocks.js',
-			array( 'wp-blocks', 'wp-i18n', 'wp-element', 'wp-components', 'wp-editor', 'wp-edit-post', 'wp-plugins' ),
-			LSX_BLOCKS_VER
-		);
-	}
-}
